read_lora/requirements.txt
pyserial
prometheus-client
influxdb-client
